Now getting data when loading the main activity instead of the feed fragment
