+++ app/src/main/java/com/victoryroad/cheers/MyFeedFragment.java
@@ -40,7 +40,7 @@ public class MyFeedFragment extends Fragment {
     // TODO: Customize parameters
     private int mColumnCount = 1;
     private OnListFragmentInteractionListener mListener;
-    public MyDrinkCardRecyclerViewAdapter mAdapter;
+    public static MyDrinkCardRecyclerViewAdapter mAdapter;
     public static List<CheckIn> CheckIns = new ArrayList<>();
 
     /**
@@ -70,8 +70,6 @@ public class MyFeedFragment extends Fragment {
 
         mAdapter = new MyDrinkCardRecyclerViewAdapter(CheckIns, mListener);
 
-        getDrinksForCurrentUser();
-
 //        CheckIn dummy = new CheckIn("Rogue Dead Guy", new LatLng(49.314, 92.729), new Date());
 //        CheckIn dummy1 = new CheckIn("Sam Adams Boston Lager", new LatLng(49.314, 92.729), new Date());
 //        CheckIn dummy2 = new CheckIn("Margarita", new LatLng(49.314, 92.729), new Date());
@@ -85,70 +83,7 @@ public class MyFeedFragment extends Fragment {
 //        CheckIns.add(dummy4);
     }
 
-    private void getDrinksForCurrentUser() {
-        String userId = MainActivity.user.getUserID();
-        DatabaseReference ref = FirebaseDatabase.getInstance().getReference("Users").child(userId).child("Checkins");
 
-        ref.addValueEventListener(new ValueEventListener() {
-            @Override
-            public void onDataChange(DataSnapshot dataSnapshot) {
-                Iterator iter = dataSnapshot.getChildren().iterator();
-                while (iter.hasNext()) {
-                    DataSnapshot child = (DataSnapshot) iter.next();
-                    final String checkinKey = child.getKey();
-
-                    DatabaseReference checkinRef = FirebaseDatabase.getInstance().getReference("Checkins").child(checkinKey);
-                    checkinRef.addValueEventListener(new ValueEventListener() {
-                        @Override
-                        public void onDataChange(final DataSnapshot checkinDataSnapshot) {
-//                            CheckIn in = dataSnapshot.child(checkinKey).getValue(CheckIn.class);
-//                            CheckIn in = (new Gson()).fromJson(dataSnapshot.child(checkinKey), CheckIn.class);
-                            String drinkKey = checkinDataSnapshot.child("DrinkKey").getValue(String.class);
-
-                            FirebaseDatabase.getInstance().getReference("Drinks").child(drinkKey).addValueEventListener(new ValueEventListener() {
-                                @Override
-                                public void onDataChange(DataSnapshot dataSnapshot) {
-                                    String drinkName = dataSnapshot.child("Name").getValue(String.class);
-                                    ArrayList<String> categories = new ArrayList<>();
-
-                                    for (DataSnapshot category : dataSnapshot.child("Categories").getChildren()) {
-                                        categories.add(category.getKey());
-                                    }
-
-                                    double lat = checkinDataSnapshot.child("Location").child("latitude").getValue(double.class);
-                                    double lng = checkinDataSnapshot.child("Location").child("longitude").getValue(double.class);
-                                    LatLng location = new LatLng(lat, lng);
-                                    Date time = checkinDataSnapshot.child("Time").getValue(Date.class);
-//                            Date time = (new Gson()).fromJson(timeString, Date.class);
-
-                                    CheckIn checkin = new CheckIn(drinkName, location, time);
-                                    checkin.Categories = categories;
-                                    CheckIns.add(checkin);
-
-                                    mAdapter.notifyDataSetChanged();
-                                }
-
-                                @Override
-                                public void onCancelled(DatabaseError databaseError) {
-
-                                }
-                            });
-                        }
-
-                        @Override
-                        public void onCancelled(DatabaseError databaseError) {
-
-                        }
-                    });
-                }
-            }
-
-            @Override
-            public void onCancelled(DatabaseError databaseError) {
-
-            }
-        });
-    }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
